Moved routes out of components
