+++ shared/routes/community-routes/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Redirect, IndexRoute } from 'react-router';
 
-import Layout from '../Layout';
+import Layout from '../../components/Layout';
 import Community from '../../containers/Community';
 
 export default (
