+++ shared/routes/conference-routes/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Route, Redirect, IndexRoute } from 'react-router';
 
-import Layout from '../Layout';
-import Conference from '../Conference';
+import Layout from '../../components/Layout';
+import Conference from '../../components/Conference';
 
 export default (
   <Route path="/" component={Layout}>
